Refactor the API for better structure and readability. This includes breaking down the websocket module into smaller files, using dependency injection for settings, and fixing a bug in the API key verification.
