+++ api/websocket/messaging.py
@@ -1,0 +1,76 @@
+import asyncio
+import base64
+import json
+from google.genai.types import (
+    Part,
+    Content,
+    Blob,
+)
+
+
+async def agent_to_client_messaging(websocket, live_events):
+    """Agent to client communication"""
+    while True:
+        async for event in live_events:
+
+            # If the turn complete or interrupted, send it
+            if event.turn_complete or event.interrupted:
+                message = {
+                    "turn_complete": event.turn_complete,
+                    "interrupted": event.interrupted,
+                }
+                await websocket.send_text(json.dumps(message))
+                print(f"[AGENT TO CLIENT]: {message}")
+                continue
+
+            # Read the Content and its first Part
+            part: Part = (
+                event.content and event.content.parts and event.content.parts[0]
+            )
+            if not part:
+                continue
+
+            # If it's audio, send Base64 encoded audio data
+            is_audio = part.inline_data and part.inline_data.mime_type.startswith("audio/pcm")
+            if is_audio:
+                audio_data = part.inline_data and part.inline_data.data
+                if audio_data:
+                    message = {
+                        "mime_type": "audio/pcm",
+                        "data": base64.b64encode(audio_data).decode("ascii")
+                    }
+                    await websocket.send_text(json.dumps(message))
+                    print(f"[AGENT TO CLIENT]: audio/pcm: {len(audio_data)} bytes.")
+                    continue
+
+            # If it's text and a parial text, send it
+            if part.text and event.partial:
+                message = {
+                    "mime_type": "text/plain",
+                    "data": part.text
+                }
+                await websocket.send_text(json.dumps(message))
+                print(f"[AGENT TO CLIENT]: text/plain: {message}")
+
+
+async def client_to_agent_messaging(websocket, live_request_queue):
+    """Client to agent communication"""
+    while True:
+        # Decode JSON message
+        message_json = await websocket.receive_text()
+        message = json.loads(message_json)
+        mime_type = message["mime_type"]
+        data = message["data"]
+
+        # Send the message to the agent
+        if mime_type == "text/plain":
+            # Send a text message
+            content = Content(role="user", parts=[Part.from_text(text=data)])
+            live_request_queue.send_content(content=content)
+            print(f"[CLIENT TO AGENT]: {data}")
+        elif mime_type == "audio/pcm":
+            # Send an audio data
+            decoded_data = base64.b64decode(data)
+            live_request_queue.send_realtime(Blob(data=decoded_data, mime_type=mime_type))
+        else:
+            raise ValueError(f"Mime type not supported: {mime_type}")
